Implement linear dimension scale factor
Implement support for user-adjustable linear dimension scale factor.
Setting it to 1.0 means that dimensions are scaled in the same
way as distances, with one AU. Setting it at higher values enlarges
the celestial bodies, but keeps their size proportions.

Add a "dimension scaling" slider to the UI, which allows user
to adjust the scaling from 1x to 50x.