Move trajectory tracker into physics engine
Move trajectory tracker functionality into physics engine part, and
have it updated on every internal discrete simulation step.

This should result in smoother trajectories when update step is
relatively large, but the internal simulation step remains relatively
fine-grained (for example, with slider set to 100 days/sec, and
internal simulation step fixed to 3600 sec). The change is
noticeable especially with Mercury and Moon.

It does not help that much with auto-scaled internal simulation step
with large update steps, because there the internal simulation
step itself is also quite large; for example, with slider set to
100 days/sec, the internal simulation step ends up being auto-set
to 24 hours, which means that the Moon's trajectory is still quite
jagged.