CelestialBodyView: split Update() method into two parts
Split `CelestialBody.Update()` method into two parts.

The first part (`UpdatePhysicalProperties`) updates only the properties
that depend solely on physical properties (e.g., position, orientation,
trail). This is essentially the codepath that was previously executed
only if `Update` was called with `dataChanged=true` argument.

The second part (`UpdateVisualization`) performs dynamic scaling
of celestial body's size, and hiding/showing of the children (i.e.,
moons).

This split is necessary because the updates in the second part depend
on camera properties (position and orientation); and if we are tracking
a celestial body, the camera can be updated only after visualization's
position is updated from the physical properties...

In the old monolithic `Update` method, this meant that dynamic scaling
was performed with camera position from previous update, which would
turn out to be grossly incorrect with very large update steps (and
for planets with shorter orbits).