CelestialBodyView: rework positioning of the name scene node
Have the code for positioning of the celestial body's name scene
node re-use the body-to-camera distance that was pre-computed and
stored in the calling codepath.

The previous approach used text-node-to-camera distance, but that
seems to have introduced a cyclic dependency loop: i.e., the distance
was computed from old text-node position and camera's position, and
then used to compute the new text-node position. This worked for
small changes, but now that the camera is updated before text-node
is (the split from previous commit), we can see large oscillations
in text node size estimated from this distance when tracking a
planet with large update step setting.