Allow scenarios to specify default camera distance
Allow scenarios to specify default camera distance; this is used
to reset the view in cases when scenario does not specify the
celestial body on which the view should be centered by default.
In such cases, the view is reset to point at (0, 0, 0) with the
supplied distance value (which should be in display units).