CelestialBodyView: properly orient and shift the orbit ellipse
The orbit ellipse needs to be both rotated (using longitude of
ascending node and argumet of periapsis in addition to inclination)
and shifted (so that the parent celestial body is moved into one
of the foci).

Our orbit ellipses now seem properly aligned with the planet positions
obtained from the solar system almanac.