CelestialBodyView: apply orbital inclination to body's sphere
Attempt to correctly transform the celestial body's sphere according
to both orbital inclination and axial tilt.

The main difficulty here is that in contrast to, for example, the
orbital ellipse node, which we transform by specifying the direction
vectors, the sphere needs to be transformed using the Transform
property (and, for example the Y axis in sphere's coordinate system
is assumed to point upwards).

To make implementation and debugging easier, we add axis visualizations
(via `AxisLineNode` that shares Transform with the `SphereModelNode`)
for both ecliptic axes (i.e., to show that the axes and the sphere are
properly inclined; the sphere's Y axis should be perpendicular to the
orbital/ecliptic plane) and for the rotational axes (to show that the
sphere is tilted w.r.t. the ecliptic plane). The rotation of rotational
axes w.r.t. ecliptic axes also shows what the body's rotation plane is.